requirements.txt
et_xmlfile==2.0.0
greenlet==3.1.1
numpy==2.2.2
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pillow==11.1.0
playwright==1.50.0
pyee==12.1.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
pytz==2025.1
six==1.17.0
typing_extensions==4.12.2
tzdata==2025.1
*.txt
*.log